pyproject.toml
[tool.poetry]
name = "project-with-stripe"
version = "0.1.0"
description = ""
authors = ["arturkalimullin <noti1996@gmail.com>"]
readme = "README.md"
packages = [{include = "project_with_stripe"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
stripe = "^5.0.0"
djangorestframework = "^3.14.0"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
